search-site/requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
rank_bm25>=0.2.2
scikit-learn>=1.3.0
gunicorn>=21.2.0
azure-search-documents>=11.6.0
azure-identity>=1.25.0
openai>=1.0.0
python-dotenv>=1.0.0
